requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
openpyxl
python-dotenv
pydantic
pydantic-settings
pandera
sqlite-utils
typer

